Fixes bug with refsections
Once you've used a refsection you need to close it if you don't want to
use a refsection for later parts.  This is most easily accomplished by
starting a new refsection when printing the list of references.

With this construction, \references prints a list of the references used
since the previous \references command (or use of \newrefsection).